dataset-creator-function/requirements.txt
functions-framework==3.*
google-cloud-storage==3.3.0
google-cloud-firestore==2.21.0
google-genai==1.32.0